completions: reformat and change bindings
There MIGHT be some functional changes in here that makes jdtls
completion work - not sure. I do know that we make the Tab and Shift+Tab
bindings not actually preview the snippet, since it was getting
confusing as it doesn't preview AFTER the cursor, only before. Now, you
use Tab and Shift+Tab to choose one, and Enter to complete it.