plugins: add bufferline and yazi
Perhaps the Neovim people say you shouldn't have a tab bar, but would
it's either I use the Neovim tab bar, the Yazi tab bar, or the Kitty tab
bar. So, I'll be sticking firm on this one. I am getting a strange
behavior where Yazi opens everything in new buffers, but that's a fix
for another day.