bufferline: use tabs by default
Tabs and buffers both have their pros and cons. But, tabs make it super
simple to open help, they make quitting them less scary since you have
to :q all of them, and they allow duplicates better. The only thing I
miss about buffers is that yazi.nvim has a bug with them - and it's a
bug I like. It opens everything in a new buffer by default, which is
great for preventing me keeping everything in the same tab. The only
way I can find to replicate it currently is to bind `o` in yazi.nvim to
open a tab - and that causes issues when using zoxide, since I can no
longer type in `o`. Maybe there's another way with the internal commands
of yazi.nvim - but that's a commit for another day.