plugins/yazi: make `tt` open new tab from Yazi
Helping to accomplish the behavior I lost from switching to tabs, not
buffers. The old behavior works if you want to replace your tab, but
this basically just grabs the open_file_in_tab function from yazi.nvim,
and has it use it for all files. It's not perfect - I wish it would show
a blank page in the background, but I can't do that without losing the
nice Yazi behavior where it'll create a new tab in the directory of the
current tab. Unless I find a way to replace that with Lua stuff, this is
fine.