plugins: add gitsigns and wilder
We use a wilder fork with some relevant bug fixes. We also add gitsigns.
Finally, we reorganize the custom file so we'll never forget to add
something to the exported list - instead using a wrapping attrset and
using `attrValues` to grab all the values. Better than lists, since
there everything has to be parenthesized.