packages/plugins: collect from folders
We add `llakaLib` as a dependency for `collectDirectoryPackages`. We
don't have any self-reliance right now, so technically
packagesFromDirectoryRecursive would do the job, but it lets me add it
eventually. We needed to add a package to the `packages` folder for nix
to not error, so I added pylsp from the Helix config. We copy it over
mostly unchanged, but since we have 24.11 on here, we had to disable one
more test. Seems to be working, though - I did some basic comparison to
Helix and didn't see any differences in behavior.