commands: init
We get off to a strong start with my first command: a way to quickly
turn brackets into Allman style (aka the best style). It would've been
nicer to have this kind of thing done automatically - and there are ways
to do that, I starred a Prettier extension on Github that did this,
ESLint has a setting, and there's other people that have done it with
lh-brackets (although I wasn't able to get that working). But this
solution is agnostic to any language, and it DOES work on everything
I've tested it on!