outputs: rewrite to export packages/plugins
Rather than grabbing the packages and plugins lower-down in the
`packages.nix` and `plugins.nix` files, we instead grab them up top at
the flake level. We then export them as flake outputs, for easy testing
in the repl or wherever you do your testing.

We don't export the non-custom plugins and packages. This is because to
export them, they would need to be in the form of an attrset. It's very
easy to turn an attrset into a list - but not so the other way around.
If we didn't want to do the type change, we would need to write
`packages.nix` and `plugins.nix` to instead export attrsets - which is
possible with `inherit` syntax, but it's much less clean than a simple
list. The best solution to this would be *set slicing*. Lix, save me!
https://wiki.lix.systems/books/lix-contributors/page/nix-lang-v2

Anyways, to avoid these frustrations, we only export the custom stuff.
This is fine - it keeps the flake outputs less busy, anyways.