plugins/pairs: disable auto parentheses in gleam
This was giving a false positive, where if I did:
```
3 |> io.deb
```
and pressed Tab to complete, it would add parentheses, when it didn't
need to, since we were in a pipe.

Hopefully this won't make auto pairs start not working in other
languages, since I technically didn't provide the `*` case. I assume
it's smart enough to know what I want.