options: fix comment continuation
Fixes #1. My best guess for the  cause was that the comment continuation
must have been coming from tree-sitter. When a tree-sitter grammar
wasn't providing something to continue the comments, it wouldn't have
anything to fall back on. We now enable comment continuation in
formatexpr, so it works based off the value of the commentstring.