conform: init
We like autoformatting in *most* languages. Conform gives me the
wonderful thing of an easy way to pass shell commands in to format a
language, if the LSP doesn't provide one.

There are two kinds of files I *don't* want formatted. Either a filetype
doesn't have any good formatters, or it's someone else's file, and you
don't want to be messing with the formatting. We add conditions to
handle both of these. First, if you put `vim.b.disable_autoformat =
true` in your language's ftplugin, it'll be disabled. Second, if your
file is in `Documents/repos/*`, it won't be formatted. That's where I
store all my random repos, so looking good!

We also add `:fmt` for when I need to work around one of these. The gods
above say you shouldn't put a lowercase custom command - which I get.
But I'm not making a lowercase custom command. I'm making an uppercase
custom command, and abbreviating the lowercase version to the uppercase
version. Listen, I have a lot of issues with the Helix maintainers, but
at least they understood that :fmt is necessary.