session: only create one session for a git repo
Some wonderful UX here! Sometimes, I'll cd into a subdirectory of a git
repo. auto-session ends up creating a new session for that subdir, which
typically isn't what I want. To get around this, we only *create* a
session if we're at the root of a git repo. If we're in a git repo, but
not at the root, we'll reuse the session from the root. This means I can
be cd'ed anywhere in the repo, and still get the same session!