cmp: replace with `blink`
Lots of stuff in this. First, we replace almost everything that nvim-cmp
did with the blink equivalent. The only thing I don't have set up yet is
completion for github issues and PRs. We also replace a few other
plugins - we're no longer using wilder-nvim, for example, instead opting
for the blink equivalent. We also aren't using lsp-signature - I had no
issues with it, but I like the blink one more. So far, blink is
wonderful - it's very batteries included in a way I like.