snacks: switch tabs if the buffer is already open
Wish I could've done this without copying in the function from the
snacks source, but I don't think there's a better way here. Now, the
`snacks_new_tab` function opens in a new tab if the buffer isn't already
open. If the buffer *is* open, it'll just swap tabs! Turns out this
functionality was already in snacks, but it was being mean and would
only do one of these. Now, it does both!