structure: rewrite
Hopefully, this makes things more readable! We get rid of the
`flake.nix` and `outputs.nix` separation, instead choosing to have a
single `flake.nix` for handling everything. I like this separation in
theory, but I want accessibility here.

We also now have all the custom derivations folders in their own folder.
I'd prefer to name this something like `nix` or `extras`, but I need it
to come after `nvim` in the alphabet for my life with Yazi cd'ing into
the folder. I also work on the comments in the flake.nix, so it's
hopefully clearer.