mini-ai: add custom comment textobject
With a day's work of regex, and finally a solution that was much simpler
and faster, we've done it! My magnum opus. We don't just have custom
`ic` and `ac` logic, for keeping vs removing the beginning of the
textobject - but we've also got EOL comma support, which `mini.comment`
didn't have before!

And now, I sleep.