mini-indentscope: don't include border
I need this to be dependable for `cii` or `dai`, and currently it's not.

In the future, I also want to try out `snacks.scope`, since it seems to
have better support for something I really want in my indent scope: good
"sibling" support. Basically, the textobject should start from my
current line, and end when the indentation RETURNS to the level I'm
currently at.