keybinds: `H` -> `!`
H and L are actually useful, and I'd like to use them more for
navigating around. I never really use L, since $ is ergonomic enough -
but ^ sucks to reach. ! is super easy to reach, and it doesn't have any
interesting functionality - it just pipes the selected text into a shell
command, but you can replicate it with `v{motion}:!`.