options: don't use cindent
`indentexpr` always overrides cindent if it exists for the current
language. But if it DOESN'T
exist, cindent isn't a good default - it has terrible handling for stuff
like `#` comments, and just does a lot of strange things. So, we assume
that our language will provide an indentexpr, and remove the cindent
fallback.