surround: change bindings
Swap forward and backwards brackets, and set some `cur` and `line`
bindings. Now, `s` is for generic surround, and `ss` is for the full
line, just like `cc` and `dd`. `S` is for if you want the surround to be
on newlines.

Note that we keep insert mode and `change_line` maps unbound. I never
do control binds in insert mode, and `change_line` is weird - changing
something shouldn't mess with whether it's on newlines or not.