fzf-lua: use faux normal mode for most pickers
Normal mode is great for some things like the buffer picker, but the
double escape sucks for most pickers. To get around this, we use alt
bindings for `hjkl` navigation. This is preferable to a full mode shift,
and will hopefully keep me fuzzy picking for longer.

All the normal mode logic is now moved to the `buffers` picker, where it
can be centralized. If we ever get a new picker that only has a few
elements, this could be used for that too.

Finally, we swap the accept bind from `o` to `l`. I find this more
intuitive so far - maybe I'll revert it later.