flake: remove llakaLib dependency
I want to improve the eval time of this as much as I can. To start out,
we stop using the `collectDirectoryPackages` wrapper, and instead use a
simple locally-defined wrapper. It provides the important thing -
self-reference - even if it does lose automatic updates based on the
state of the folder.