autocmds/scrolloff: init
I added this a long time ago, but ran into an issue where the `fFtT`
highlights weren't working. Turns out, the problem was that `zz` clears
highlights, even if the cursor's position doesn't change. We work around
this here by detecting whether the current line is already centered, and
only running `zz` if it's not. See the comments for more details!