mini: use single comment textobject, differentiated by operator
I saw someone else's textobject using `vim.v.operator` for setting
different behavior depending on the mode. In theory, textobjects aren't
supposed to do this - the different operators are just choosing what
happens with the selected text. But I think having a single textobject
here has more benefits than it does drabacks.

There are two categories of things I do with comments:
- change it
- delete it / something else

When changing a commentt, I want to keep the comment header, so I can
add some new content. When doing something else, I want the header
included. Before, I implemented this through two separate textobjects -
you could do `cic` to "change inside comment", and `dac` to "delete
around comment". But you'd only ever use the `ic` textobject for changng
a comment - so why not just unify them to one? Thanks to `vim.operator`,
we can do this.

In the new system, we just have `ic`. `cic` will keep the header, while
`dic`, `vic`, `gwic`, etc. will remove it. This hypothetically breaks
the unwritten textobject rules, as you're not supposed to do `i` version
without an `a` equivalent.

As part of this, we moved the implemention to not depend on `mini.ai`. I
didn't actually have anything that required it, and using the vanilla
mechanisms means there's less things that can break.

Finally, we did an extra bit of conditional logic for EOL comments. If
we only have two categories - change and non-change - then `vic` would
include the initial space between the code and the comment. This isn't
helpful if I just want to copy a comment, so we only include that
initial space when using the `d` operator.