various: speed up cwd access
Rather than running the relevant `git` commands, we use `vim.fs.root`.
We also only run the command once, rather than putting it in a function,
so it doesn't have to be recomputed. Getting a meaningful improvement to
startuptime from this - about 50ms when closing and reopening neovim
with a few files open.