oil: use ctrl bindings, add wraparound search
`ctrl+f` to search forward from your current location - and `ctrl+;` to
repeat the previous search. We always start the search one ahead of your
current location, so if you're on the first file starting with `g`,
doing `<C-f>g` will bring you to the next one - and when you're out of
g-files, you'll wrap around.