oil: use normal search, add highlighting
My original impl might be cool, but I was able to find a way to search a
specific column with normal search: https://stackoverflow.com/a/39749505
got me on the right track. We also highlight all lines matching the
pattern. I'm sure this could be optimized, so we only highlighted lines
currently displayed on the screen - but I'm happy with this impl for
now.