ftplugin/nix/pairs: rework `;` logic
I like the new way things are done. Before, typing `;` if it was already
there would mean it would just "type over" the existing semicolon. This
was specifically annoying if you tried to type something like:

```
{
  x = with pkgs; [ git ];
}
```

The first semicolon would be auto-inserted on `= `, and then when you
tried to add one yourself for `with`, it would overwrite the existing
one, so you'd have one when you needed two.

Now, if there's already a semicolon after your cursor, it won't add
another one - but if you manually type a semicolon, it'll actually be
entered.

We also add auto-semicolon for `inherit`, and prevent `!= ` from
triggering a semicolon. I tried to do the same for `==`, but wasn't able
to get it to work.